+++ com.unity.render-pipelines.core/Runtime/Common/CommonStructs.cs
@@ -2,7 +2,6 @@ using System;
 
 namespace UnityEngine.Rendering
 {
-    #if !UNITY_2021_2_OR_NEWER
     /// <summary>
     /// Render Textures clear flag.
     /// </summary>
@@ -15,9 +14,10 @@ namespace UnityEngine.Rendering
         Color = 1,
         /// <summary>Clear the depth buffer.</summary>
         Depth = 2,
+        /// <summary>Clear the stencil buffer.</summary>
+        Stencil = 4,
 
-        /// <summary>Clear both color and depth buffers.</summary>
-        All = Depth | Color
+        /// <summary>Clear both color, depth and stencil buffers.</summary>
+        All = Color | Depth | Stencil
     }
-    #endif
 }