+++ com.unity.render-pipelines.core/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [12.0.0] - 2021-01-11
 
 ### Changed
-- ClearFlag moved into the Unity Core module (Unity 2021.2).
+- ClearFlag.Depth does not implicitely clear stencil anymore. ClearFlag.Stencil added.
 
 Version Updated
 The version number for this package has increased due to a version update of a related graphics package.